fix: moves config.go back to top-level package config
